Extract lang tasks and don't zip
